feat(DEQ-99): unify Add Stack and Edit Stack views
Create StackEditorView that handles both create and edit modes, enabling
users to add reminders when creating a new Stack (previously impossible).

Key changes:
- New StackEditorView with Mode enum (.create, .edit(Stack))
- Create mode auto-saves draft on first title input
- Reminders section appears in create mode when draft exists
- Draft stacks use create-mode UI for consistent experience
- Split into extensions to meet SwiftLint file length limits

Files replaced:
- AddStackView.swift (224 lines) → StackEditorView
- StackDetailView.swift (497 lines) → StackEditorView

Call site updates:
- MainTabView (iOS + macOS)
- HomeView
- DraftsView
- CompletedStacksView (with isReadOnly: true)
